refactor: Align ApiProblem class naming with HTTP Status codes (#96)
Using the `ApiProblem` became a little bit cumbersome. At first I
thought it is a good idea to have a different naming from the HTTP
status codes but the translation between those two different terms
representing the same response code was getting a bit clunky.

Rename the existing `ApiProblem`s to be aligned with the HTTP status
code and move the classes into the object.
Replace `implicit val` for codecs and schemas with Scala 3 `given`.